refactor(WIP): merge apps
