+++ bcs-app/backend/apps/constants.py
@@ -13,13 +13,43 @@
 #
 """容器管理使用常量
 """
+import logging
 from enum import Enum
 
-from django.conf import settings
-
-from backend.apps.constants_bk import *  # noqa
-from backend.apps.constants_bk import ProjectKind
 from backend.utils.basic import ChoicesEnum
+
+logger = logging.getLogger(__name__)
+
+# 不检查IP是否重复的业务
+SKIP_BIZ_INFO = {}
+
+# 白名单接入的业务
+BIND_BIZ_ID = []
+BIND_BIZ_ID_USER = {}
+
+# CC MODULE INFO
+CC_MODULE_INFO = {}
+
+# k8s 平台服务用的集群
+K8S_PLAT_CLUSTER_ID = []
+
+# master in binded biz
+BCS_APP_ID = ""
+
+# verify resource code for perm
+verify_resource_exist = False
+
+# nginx ingress controller path
+CONTROLLER_IMAGE_PATH = "public/bcs/k8s/nginx-ingress-controller"
+BACKEND_IMAGE_PATH = "public/bcs/k8s/defaultbackend"
+
+
+# 项目类型
+class ProjectKind(ChoicesEnum):
+    K8S = 1
+    MESOS = 2
+
+    _choices_labels = ((1, 'Kubernetes'), (2, 'Mesos'))
 
 
 class NodeStatus(Enum):
@@ -99,3 +129,9 @@ ALL_CLUSTER_FLAG = "ALL"
 
 # 敏感单词
 SENSITIVE_KEYWORD = ["access_token", "bk_app_secret", "X-BKAPI-AUTHORIZATION", "X-BK-APP-SECRET"]
+
+
+try:
+    from .constants_ext import *  # noqa
+except ImportError as e:
+    logger.debug('Load extension failed: %s', e)
